LG-88: Move scripts to folder
