+++ src/components/personal/PersonalCabinet.vue
@@ -65,7 +65,7 @@
 
 <script>
 
-import AuthDataService from "@/auth/AuthDataService";
+import AuthDataService from "@/scripts/auth/AuthDataService";
 import {useAuthStore} from "@/stores/authStore";
 
 export default {
